Individual Novels Extraction

This function allows you to extract the novel of your choice based on the title you choose along with any extra parameters you'd like to see with it such as the year or the author.

Usage

polish(..., vars)

Arguments

...

Input any title from the novels() function

vars

Input any column variable of the novels() function

Value

A dataframe with title, text, and then any column variable of your choice

Examples

polish(title == "Frankenstein", vars = "year")
#> # A tibble: 11,082 × 3
#> # Groups:   title, text [10,266]
#>    title        text                                        year
#>    <chr>        <chr>                                      <dbl>
#>  1 Frankenstein FRANKENSTEIN                                1818
#>  2 Frankenstein OR                                          1818
#>  3 Frankenstein THE MODERN PROMETHEUS                       1818
#>  4 Frankenstein BY MARY W                                   1818
#>  5 Frankenstein SHELLEY                                     1818
#>  6 Frankenstein PREFACE                                     1818
#>  7 Frankenstein The event on which this fiction is founded  1818
#>  8 Frankenstein has been supposed                           1818
#>  9 Frankenstein by Dr                                       1818
#> 10 Frankenstein Darwin                                      1818
#> # ℹ 11,072 more rows